+++ assign_v2/boost/assign/v2/ref/aux_/list/holder/tail.hpp
@@ -12,7 +12,7 @@
 #include <boost/assign/v2/support/keyword.hpp>
 #include <boost/assign/v2/ref/aux_/list/fwd.hpp>
 #include <boost/config.hpp>
-#include <boost/mpl/int.hpp>
+#include <boost/mpl/size_t.hpp>
 #include <boost/mpl/plus.hpp>
 #include <boost/call_traits.hpp>
 
@@ -28,7 +28,7 @@ namespace list_aux{
         typedef T tail_type;
         typedef ::boost::mpl::plus<
             typename T::tail_static_size,
-            ::boost::mpl::int_<1>
+            ::boost::mpl::size_t<1>
         > tail_static_size;
         typedef T const& result_of_tail_type;
 
@@ -46,7 +46,7 @@ namespace list_aux{
     struct tail_holder<nil_>
     {
         typedef nil_ tail_type;
-        typedef ::boost::mpl::int_<0> tail_static_size;
+        typedef ::boost::mpl::size_t<0> tail_static_size;
         tail_holder(){}
 
         typedef tail_type result_of_tail_type;
